Added level transitions
levels are on a grid of 8 by 8, starting screen is at 8,1. When you transition at a side of the screen, the level gets unloaded (which can be majorly optimized!!) and the new level gets loaded.
LevelGrid.pdn is a Pain.net project file to help with the layout of individual levels. Just name a layer of a level to X_Y_Layer_1/2/3 to make it automatically load when a player enters that space.

A present bug is the player is always already rendered in a new level so he disapears to the background, maybe re render the players graphic when loading a level?

Leveltransition blocks are transparant ingame but can be painted in layer 2 with colors 0,255,255.